# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Board of Issues task management system.",
        "title": "Board Of Issues API",
        "contact": {
            "name": "Nikita Kleymenov",
            "url": "https://t.me/n1ktarchik",
            "email": "klejmenov663@email.com"
        },
        "version": "2.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tasks": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth,IdempotencyKey": []
                    }
                ],
                "description": "Create a new task in a specific desk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create a task",
                "parameters": [
                    {
                        "description": "Task Info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.TaskRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Task",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.Task"
                        }
                    },
                    "400": {
                        "description": "Possible: task_name_too_short, invalid_desk_id",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "desk_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update task name, description or deadline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "description": "New task data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.UpdateTaskRequestDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated task",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.Task"
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_deadline_format, task_name_too_short",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "not_a_desk_member",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "task_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/complete": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mark a specific task as done",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Complete task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "\"550e8400-e29b-41d4-a716-446655440000\"",
                        "description": "TASK ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: task with ID ... has been completed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_uuid, bad_request",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "not_a_desk_member",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "task_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "task_already_completed",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "N1ktarchik_Board_of_issues_internal_core_domain.Task": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "636e856-e12b-56d9-f987-333222561234"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "deadline": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Task description"
                },
                "desk_id": {
                    "type": "string",
                    "example": "832t758-a12g-47y9-i999-123456789098"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "example": "Task name"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "error": {
                    "type": "string",
                    "example": "any string-code of error"
                },
                "message": {
                    "type": "string",
                    "example": "any error message"
                }
            }
        },
...
        "internal_features_tasks_transport_http.TaskRequestDTO": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "636e856-e12b-56d9-f987-333222561234"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "deadline": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Task description"
                },
                "desk_id": {
                    "type": "string",
                    "example": "832t758-a12g-47y9-i999-123456789098"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "example": "Task name"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "internal_features_tasks_transport_http.UpdateTaskRequestDTO": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "636e856-e12b-56d9-f987-333222561234"
                },
                "deadline": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Task description"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "example": "Task name"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token in the format: Bearer \u003cJWT_TOKEN\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "IdempotencyKey": {
            "description": "An idempotent key to prevent duplicates (UUID)",
            "type": "apiKey",
            "name": "X-Req-Key",
            "in": "header"
        }
    }
}